Add Code component
